Refactors three primary functions to their own files
